Refine GPU admission and saved-state planning; record ARM qualification
Remove ambient shim GPU grants and translate admitted Docker device requests with bounded render-node permissions. Add profile-specific PC VirGL2 worker wiring while retaining qualification gates, renew native Mac saved-state plans, and fix PSCI affinity argument validation.

Record ARM installer, update, recovery and cold-snapshot development evidence plus verified container guest builds. GPU compute, production lifecycle qualification and direct native Mac mutable-disk integration coverage remain open in PLAN.md. Exclude unfinished x86 Mesa producer scripts.

+++ Packages/ContainerizationEngine/Sources/DoryHV/ARMPSCICPUState.swift
@@ -44,7 +44,7 @@ struct ARMPSCICPUState {
         states[index] = .on
     }
 
-    func affinityInfo(target: UInt64, lowestLevel: UInt32) -> Int64 {
+    func affinityInfo(target: UInt64, lowestLevel: UInt64) -> Int64 {
         // PSCI 1.0 permits an implementation to support only affinity level zero.
         guard lowestLevel == 0, let index = index(for: target) else { return -2 }
         return states[index].rawValue
